cell_tracking/requirements.txt
# Confirmed to work with Python 3.7.9
# MUST INSTALL Cython FIRST (ripser setup/wheel error)
imageio==2.9.0
matplotlib==3.3.0
numpy==1.19.0
pandas==1.1.1
persim==0.1.3
Pillow==7.2.0
ripser==0.5.3
seaborn==0.11.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.1
git+https://github.com/joferkington/mpldatacursor@2df44a3912c2684b2d66fadc7cacbb9f60a15186#egg=mpldatacursor
